requirements.txt
psd-tools>=1.10.0
Pillow>=10.0.0
click>=8.0.0
tqdm>=4.64.0
colorama>=0.4.4